# swagger.yaml
openapi: 3.0.0
info:
  title: Anime Api
  contact: {}
  version: "1.0"
servers:
  - url: https://anime-api-aiup.onrender.com/api/v1
    variables: {}
paths:
...
  /user/favourites:
    get:
      tags:
        - User
      summary: Get Favourites
      operationId: GetFavourites
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    post:
      tags:
        - User
      summary: Add To Favourites
      operationId: AddToFavourites
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/AddToFavouritesRequest"
                - example:
                    animeId: 641a172d04e2678209afa4e8
            example:
              animeId: 641a172d04e2678209afa4e8
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
components:
  schemas:
...
    AddToFavouritesRequest:
      title: AddToFavouritesRequest
      required:
        - animeId
      type: object
      properties:
        animeId:
          type: string
      example:
        animeId: 641a172d04e2678209afa4e8
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
  - httpBearer: []
...